Remove pybind deps from importer and annotator
