+++ python/torch_mlir/dialects/torch/importer/jit_ir/csrc/function_importer.h
@@ -13,7 +13,6 @@
 #include <memory>
 
 #include "node_importer.h"
-#include "pybind.h"
 
 #include "mlir-c/IR.h"
