+++ python/torch_mlir/dialects/torch/importer/jit_ir/csrc/get_registered_ops.h
@@ -16,6 +16,7 @@
 #define TORCHMLIRJITIRIMPORTER_CSRC_GETREGISTEREDOPS_H
 
 #include "pybind.h"
+#include <torch/csrc/utils/pybind.h>
 
 namespace torch_mlir {
