+++ python/torch_mlir/dialects/torch/importer/jit_ir/csrc/node_importer.h
@@ -12,8 +12,6 @@
 
 #include <memory>
 
-#include "pybind.h"
-
 #include "mlir-c/IR.h"
 
 #include <torch/csrc/jit/api/compilation_unit.h>
