+++ python/torch_mlir/dialects/torch/importer/jit_ir/csrc/pybind.h
@@ -16,15 +16,9 @@
 
 #include <torch/csrc/utils/pybind.h>
 
+namespace py = pybind11;
 namespace torch_mlir {
-
-/// Thrown on failure when details are in MLIR emitted diagnostics.
-class mlir_diagnostic_emitted : public std::runtime_error {
-public:
-  mlir_diagnostic_emitted(const char *what) : std::runtime_error(what) {}
-  mlir_diagnostic_emitted() : std::runtime_error("see diagnostics") {}
-};
-
+void initClassAnnotatorBindings(py::module &m);
 } // namespace torch_mlir
 
 #endif // TORCHMLIRJITIRIMPORTER_CSRC_PYBIND_H
